Improve payout success state
